Add `UseMauiCommunityToolkitCore`, Move `Microsoft.Maui.Controls` Out of `CommunityToolkit.Maui.Core`
