+++ samples/CommunityToolkit.Maui.Sample/MauiProgram.cs
@@ -11,8 +11,8 @@ public static class MauiProgram
 	public static MauiApp CreateMauiApp()
 	{
 		var builder = MauiApp.CreateBuilder();
-		builder.UseMauiApp<App>();
-		builder.UseCommunityToolkit();
+		builder.UseMauiApp<App>().UseMauiCommunityToolkit();
+
 		return builder.Build();
 	}
 }
