+++ src/CommunityToolkit.Maui.Core/AppBuilderExtensions.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using Microsoft.Maui.Hosting;
+
+namespace CommunityToolkit.Maui.Core;
+
+public static class AppBuilderExtensions
+{
+	public static MauiAppBuilder UseMauiCommunityToolkitCore(this MauiAppBuilder builder)
+	{
+		return builder;
+	}
+}
+
