+++ src/CommunityToolkit.Maui.Core/Primitives/SnackbarOptions.shared.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Maui;
 using Microsoft.Maui.Graphics;
 
-namespace CommunityToolkit.Maui;
+namespace CommunityToolkit.Maui.Core;
 
 /// <summary>
 /// Snackbar visual options
