+++ src/CommunityToolkit.Maui.UnitTests/Alerts/SnackBar_Tests.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using CommunityToolkit.Maui.Alerts;
+using CommunityToolkit.Maui.Core;
 using FluentAssertions;
 using Microsoft.Maui;
 using Microsoft.Maui.Controls;
