+++ src/CommunityToolkit.Maui/Alerts/Snackbar/SnackBar.shared.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
+using CommunityToolkit.Maui.Core;
 using Microsoft.Maui;
 using Microsoft.Maui.Controls;
