+++ src/CommunityToolkit.Maui/AppBuilderExtensions.cs
@@ -1,10 +1,12 @@
-﻿using Microsoft.Maui.Hosting;
+﻿using CommunityToolkit.Maui.Core;
+using Microsoft.Maui.Hosting;
 
 namespace CommunityToolkit.Maui;
+
 public static class AppBuilderExtensions
 {
-	public static MauiAppBuilder UseCommunityToolkit(this MauiAppBuilder builder)
+	public static MauiAppBuilder UseMauiCommunityToolkit(this MauiAppBuilder builder)
 	{
-		return builder;
+		return builder.UseMauiCommunityToolkitCore();
 	}
 }
