+++ src/CommunityToolkit.Maui/Behaviors/BaseBehavior.shared.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Reflection;
 using Microsoft.Maui.Controls;
 
-namespace CommunityToolkit.Maui;
+namespace CommunityToolkit.Maui.Behaviors;
 
 /// <summary>
 /// Abstract class for our behaviors to inherit.
