+++ src/CommunityToolkit.Maui/Behaviors/EventToCommandBehavior.shared.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Reflection;
 using System.Windows.Input;
+using CommunityToolkit.Maui.Converters;
 using Microsoft.Maui.Controls;
 
 namespace CommunityToolkit.Maui.Behaviors;
