+++ src/CommunityToolkit.Maui/Converters/ColorToColorConverters.shared.cs
@@ -1,3 +1,4 @@
+using CommunityToolkit.Maui.Core.Extensions;
 using CommunityToolkit.Maui.Extensions;
 using Microsoft.Maui.Graphics;
