+++ src/CommunityToolkit.Maui/Converters/ICommunityToolkitMultiValueConverter.shared.cs
@@ -2,7 +2,8 @@
 using System.Globalization;
 using Microsoft.Maui.Controls;
 
-namespace CommunityToolkit.Maui;
+namespace CommunityToolkit.Maui.Converters;
+
 /// <inheritdoc />
 public interface ICommunityToolkitMultiValueConverter : IMultiValueConverter
 {
