+++ src/CommunityToolkit.Maui/Extensions/ValueConverterExtension.shared.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using CommunityToolkit.Maui.Converters;
 using Microsoft.Maui.Controls.Xaml;
 
 namespace CommunityToolkit.Maui;
